v0.2.0 - Fixed some issues with math module - Redesigned ffw::Var to accept any data available - Removed ffw::VarObject and ffw::VarArray and replaced with better ffw::Object and ffw::Array - Redesigned a way of loading and saving JSON and XML files. Instead of LoadJson("path", &data) it is now data = LoadJson("path") - Added GetOpenGLImageTypes function - Fixed when disabling a shader, now the 2D graphics functions should work properly - Added CSV file parser - Added image loaders/writers - Added OBJ file loader - Added tests for image files - Replaced GetFilePathProperties with FileInfo structure - Added Monkey 3D example - Added GUI system

+++ include/ffw/math/vec2.inl
@@ -235,7 +235,7 @@ inline ffw::Vec2<T>& ffw::Vec2<T>::Scale(const T Val){
 //==============================================================================
 template <class T>
 inline float ffw::Vec2<T>::Length() const{
-    return sqrtf(x*x + y*y);
+    return sqrtf(float(x*x + y*y));
 }
 //==============================================================================
 template <class T>
